[PM-17563] Move Task module to libs/common/vault to avoid Angular dependency
