+++ apps/browser/src/vault/popup/components/at-risk-passwords/at-risk-passwords.component.ts
@@ -15,6 +15,8 @@ import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.servic
 import { PlatformUtilsService } from "@bitwarden/common/platform/abstractions/platform-utils.service";
 import { CipherService } from "@bitwarden/common/vault/abstractions/cipher.service";
 import { CipherView } from "@bitwarden/common/vault/models/view/cipher.view";
+import { SecurityTaskType, TaskService } from "@bitwarden/common/vault/tasks";
+import { filterOutNullish } from "@bitwarden/common/vault/utils/observable-utilities";
 import {
   BadgeModule,
   ButtonModule,
@@ -28,10 +30,7 @@ import {
 import {
   ChangeLoginPasswordService,
   DefaultChangeLoginPasswordService,
-  filterOutNullish,
   PasswordRepromptService,
-  SecurityTaskType,
-  TaskService,
   VaultCarouselModule,
 } from "@bitwarden/vault";
