+++ apps/web/src/app/auth/settings/emergency-access/view/emergency-view-dialog.component.spec.ts
@@ -17,8 +17,9 @@ import { CipherService } from "@bitwarden/common/vault/abstractions/cipher.servi
 import { FolderService } from "@bitwarden/common/vault/abstractions/folder/folder.service.abstraction";
 import { CipherType } from "@bitwarden/common/vault/enums";
 import { CipherView } from "@bitwarden/common/vault/models/view/cipher.view";
+import { TaskService } from "@bitwarden/common/vault/tasks";
 import { DialogService } from "@bitwarden/components";
-import { ChangeLoginPasswordService, TaskService } from "@bitwarden/vault";
+import { ChangeLoginPasswordService } from "@bitwarden/vault";
 
 import { EmergencyViewDialogComponent } from "./emergency-view-dialog.component";
 
@@ -55,6 +56,7 @@ describe("EmergencyViewDialogComponent", () => {
         { provide: DialogRef, useValue: { close } },
         { provide: DIALOG_DATA, useValue: { cipher: mockCipher } },
         { provide: AccountService, useValue: accountService },
+        { provide: TaskService, useValue: mock<TaskService>() },
       ],
     })
       .overrideComponent(EmergencyViewDialogComponent, {
@@ -71,10 +73,6 @@ describe("EmergencyViewDialogComponent", () => {
         },
         add: {
           providers: [
-            {
-              provide: TaskService,
-              useValue: mock<TaskService>(),
-            },
             { provide: PlatformUtilsService, useValue: mock<PlatformUtilsService>() },
             {
               provide: ChangeLoginPasswordService,
