+++ libs/common/src/vault/utils/observable-utilities.ts
